+++ lib/src/inbound/inbound.g.dart
@@ -260,7 +260,9 @@ _InboundService _$InboundServiceFromJson(Map<String, dynamic> json) =>
     _InboundService(
       id: json['id'] as String,
       name: json['name'] as String,
-      credentials: json['credentials'] as Map<String, dynamic>?,
+      credentials: const DynamicMapConverterNullable().fromJson(
+        json['credentials'],
+      ),
       externalAccountId: json['externalAccountId'] as String?,
       updateTime: const DurationOrNullConverter().fromJson(
         json['updateTime'] as num?,
@@ -290,7 +292,9 @@ Map<String, dynamic> _$InboundServiceToJson(_InboundService instance) =>
     <String, dynamic>{
       'id': instance.id,
       'name': instance.name,
-      'credentials': instance.credentials,
+      'credentials': const DynamicMapConverterNullable().toJson(
+        instance.credentials,
+      ),
       'externalAccountId': instance.externalAccountId,
       'updateTime': const DurationOrNullConverter().toJson(instance.updateTime),
       'protocol': instance.protocol?.toJson(),
@@ -306,7 +310,9 @@ _InboundServiceInput _$InboundServiceInputFromJson(Map<String, dynamic> json) =>
     _InboundServiceInput(
       id: json['id'] as String?,
       name: json['name'] as String? ?? '',
-      credentials: json['credentials'] as Map<String, dynamic>? ?? const {},
+      credentials: json['credentials'] == null
+          ? const <String, dynamic>{}
+          : const DynamicMapConverter().fromJson(json['credentials']),
       externalAccountId: json['externalAccountId'] as String?,
       protocolId: json['protocolId'] as String?,
       structure: InboundStructureInput.fromJson(
@@ -319,7 +325,7 @@ Map<String, dynamic> _$InboundServiceInputToJson(
 ) => <String, dynamic>{
   'id': instance.id,
   'name': instance.name,
-  'credentials': instance.credentials,
+  'credentials': const DynamicMapConverter().toJson(instance.credentials),
   'externalAccountId': instance.externalAccountId,
   'protocolId': instance.protocolId,
   'structure': instance.structure.toJson(),